Move hostify method
