+++ modules/common/src/UrlHostTokenResolver.php
@@ -66,4 +66,59 @@ class UrlHostTokenResolver {
     ));
   }
 
+
+  /**
+   * Substitute the host for local URLs with a custom localhost token.
+   *
+   * @param string $resourceUrl
+   *   The URL of the resource being substituted.
+   *
+   * @return string
+   *   The resource URL with the custom localhost token.
+   */
+  public static function hostify(string $resourceUrl): string {
+    // Get HTTP server public files URL and extract the host.
+    $serverPublicFilesUrl = self::getServerPublicFilesUrl();
+    $serverPublicFilesUrl = isset($serverPublicFilesUrl) ? parse_url($serverPublicFilesUrl) : NULL;
+    $serverHost = $serverPublicFilesUrl['host'] ?? \Drupal::request()->getHost();
+    // Determine whether the resource URL has the same host as this server.
+    $resourceParsedUrl = parse_url($resourceUrl);
+    if (isset($resourceParsedUrl['host']) && $resourceParsedUrl['host'] == $serverHost) {
+      // Swap out the host portion of the resource URL with the localhost token.
+      $resourceParsedUrl['host'] = UrlHostTokenResolver::TOKEN;
+      $resourceUrl = self::unparseUrl($resourceParsedUrl);
+    }
+    return $resourceUrl;
+  }
+
+  /**
+   * Private.
+   */
+  private static function unparseUrl($parsedUrl) {
+    $url = '';
+    $urlParts = [
+      'scheme',
+      'host',
+      'port',
+      'user',
+      'pass',
+      'path',
+      'query',
+      'fragment',
+    ];
+
+    foreach ($urlParts as $part) {
+      if (!isset($parsedUrl[$part])) {
+        continue;
+      }
+      $url .= ($part == "port") ? ':' : '';
+      $url .= ($part == "query") ? '?' : '';
+      $url .= ($part == "fragment") ? '#' : '';
+      $url .= $parsedUrl[$part];
+      $url .= ($part == "scheme") ? '://' : '';
+    }
+
+    return $url;
+  }
+
 }
