+++ modules/common/tests/src/Unit/UrlHostTokenResolverTest.php
@@ -19,6 +19,13 @@ use Symfony\Component\HttpFoundation\RequestStack;
 class UrlHostTokenResolverTest extends TestCase {
 
   /**
+   * HTTP host protocol and domain for testing download URL.
+   *
+   * @var string
+   */
+  public const HOST = 'http://example.com';
+
+  /**
    *
    */
   public function test() {
@@ -38,6 +45,26 @@ class UrlHostTokenResolverTest extends TestCase {
     $string = 'blahj do bla da bla ' . UrlHostTokenResolver::TOKEN . ' after token.';
     $newString = UrlHostTokenResolver::resolve($string);
     $this->assertEquals('blahj do bla da bla replacement after token.', $newString);
+  }
+
+
+  /**
+   * Test the `Referencer::hostify()` method.
+   */
+  public function testHostify(): void {
+    // Initialize `\Drupal::container`.
+    $options = (new Options())
+      ->add('stream_wrapper_manager', StreamWrapperManager::class)
+      ->index(0);
+    $container_chain = (new Chain($this))
+      ->add(Container::class, 'get', $options)
+      ->add(PublicStream::class, 'getExternalUrl', self::HOST)
+      ->add(StreamWrapperManager::class, 'getViaUri', PublicStream::class);
+    \Drupal::setContainer($container_chain->getMock());
+    // Ensure the hostify method is properly resolving the supplied URL.
+    $this->assertEquals(
+      'http://' . UrlHostTokenResolver::TOKEN . '/' . self::FILE_PATH,
+      UrlHostTokenResolver::hostify(self::HOST . '/' . self::FILE_PATH));
   }
 
 }
