+++ modules/datastore/src/Service/ResourceLocalizer.php
@@ -129,7 +129,7 @@ class ResourceLocalizer {
     $public_dir = 'file://' . $this->drupalFiles->getPublicFilesDirectory();
     $localFileDrupalUri = str_replace($public_dir, 'public://', $localFilePath);
     $localUrl = $this->drupalFiles->fileCreateUrl($localFileDrupalUri);
-    $localUrl = Referencer::hostify($localUrl);
+    $localUrl = UrlHostTokenResolver::hostify($localUrl);
 
     $new = $resource->createNewPerspective(self::LOCAL_FILE_PERSPECTIVE, $localFilePath);
